More parallelization
The initial thought was that we didn't need a ton of parallelization here because it's 9am and who is realistically going to check the website at 9am while it's populating? This is still valid, but it makes updates during potentially busy times really risky. We can parallelize things without hammering MS servers by throwing a semaphore on here, so we've done that.

Thumbnails don't use the same semaphore here since it is a separate bottleneck.