Fix search description bug
When you search for a matcha latte, the first result (at least on my machine + prod) is a result without a description. Further matches have a hit in the description, but the description is never updated, so we end up hitting an error since the text is undefined.

Now, we update every time we find a new result in case the existing one has no data.