Fix async race condition
While the cache is not populated, if two requests come in at once, we will double-populate the cache. We now have a lock per-entry in the dailyMenusByCafeId map, so that this cannot occur. We can't just set the map to promises, because we populate this cache incrementally while we discover cafe menus

There's probably still another issue: what happens if someone requests menus while we are discovering menus (aka after clearing cache)? I think we would double-append stations.