Support date-jumping via url
Date is now pulled from URL to allow sharing. Future menus are still not allowed if the setting is off, even if the link is shared. This is unlike grouping, because a user unexpectedly seeing grouping/individual cafes is not necessarily going to lose trust in the site, but a user seeing a future menu may no longer trust the site because they may not understand the implications of having the menu be in the future.