Don't fail if we have no uniqueness data
There seems to be a case where we have no uniqueness data available (I think only one day gets fetched for the week when we try to calculate uniqueness), which results in a full failure. We should still try to fix the no-uniqueness case, but let menu items get returned anyways.