Temp improvement for stretching
Directly apply object-fit: contain onto menu-item-image. This results in some gray bars around the image for things which are shrunk, so it's not ideal... but fine for now.

Seems that we are manually setting the height to a much higher value than expected in most cases. Some scaling bug somewhere