version tag header
wanted some way to enable features on the server for clients which had updated, since browsers cache JS and therefore users may see the old msdining site on first load after an update. this has been historically annoying to deal with since breaking changes can cause users to be unable to use the site without much of a good reason facing the user (and it's hard to add the relevant UX since that requires a client update)... so instead now we send a header which contains information about the known version from the client. So far we're just using this for modifiers in search results, since that throws a user-facing error if the client doesn't support it.